[CF] AUDIT-30 Создать сервис для работы с типами событий (audit_event_type) https://jira.i-novus.ru/browse/AUDIT-30

+++ audit-service/src/main/java/ru/i_novus/ms/audit/repository/predicates/EventTypePredicates.java
@@ -8,7 +8,7 @@ public final class EventTypePredicates {
     private EventTypePredicates() {
     }
 
-    public static BooleanExpression eqAuditTypeId(Integer value) {
+    public static BooleanExpression eqAuditTypeId(Short value) {
         return QEventTypeEntity.eventTypeEntity.auditTypeId.eq(value);
     }
